requirements.txt
Django==3.0.6
django-extensions==2.2.9
django-rest-swagger==2.2.0
djangorestframework==3.11.0
requests==2.23.0
